Cambio tipo de dato VLR declarado (#17)
* Subiendo ajustes de valor declarado en entrega y comentario en preliminar.

* Subiendo corrección al campo valor declarado por el tipo de dato.

+++ src/main/java/co/igb/b1ws/client/deliverynote/Document.java
@@ -1,6 +1,7 @@
 
 package co.igb.b1ws.client.deliverynote;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import javax.xml.bind.annotation.XmlAccessType;
@@ -591,7 +592,7 @@ public class Document {
     @XmlElement(name = "U_OBSERVACION")
     protected String uobservacion;
     @XmlElement(name = "U_VR_DECLARADO")
-    protected Double uvrdeclarado;
+    protected BigDecimal uvrdeclarado;
     @XmlElement(name = "U_PUERTO_EMB")
     protected String upuertoemb;
     @XmlElement(name = "U_NAVIERA")
@@ -7213,10 +7214,10 @@ public class Document {
      * 
      * @return
      *     possible object is
-     *     {@link Double }
+     *     {@link BigDecimal }
      *     
      */
-    public Double getUVRDECLARADO() {
+    public BigDecimal getUVRDECLARADO() {
         return uvrdeclarado;
     }
 
@@ -7225,10 +7226,10 @@ public class Document {
      * 
      * @param value
      *     allowed object is
-     *     {@link Double }
+     *     {@link BigDecimal }
      *     
      */
-    public void setUVRDECLARADO(Double value) {
+    public void setUVRDECLARADO(BigDecimal value) {
         this.uvrdeclarado = value;
     }
